[164]: Move the new P8eMemorializeContract endpoint into tx.proto to simplify the msg server. Move the new Empty... funcs into scope.go (along with tests into scope_test.go. Create skeleton pieces needed on the msg in msg.go.
