+++ proto/provenance/metadata/v1/p8e/p8e.proto
@@ -10,39 +10,6 @@ option java_multiple_files = true;
 // It is copied from (or modeled after) old stuff.
 // The goal is to remove this entire file (and everything defined in it) in the near future.
 
-// P8eMsg defines the immediately deprecated p8e Metadata Msg service.
-service P8eMsg {
-  // MemorializeContract records the results of a P8e contract execution as a session and set of records in a scope
-  rpc MemorializeP8eContract(P8eMsgMemorializeP8eContractRequest) returns (P8eMsgMemorializeP8eContractResponse) {
-    option deprecated = true;
-  };
-}
-
-// P8eMsgMemorializeP8eContractRequest is the input for a MemorializeP8eContract call
-message P8eMsgMemorializeP8eContractRequest {
-  option deprecated = true;
-
-  // The scope id of the object being add or modified on blockchain.
-  string scope_id = 1;
-  // The uuid of the contract execution.
-  string group_id = 2;
-  // Unique identifier for determining contract/session execution instance
-  string execution_id = 3;
-  // The new recitals for the scope.  Used in leu of Contract for direct ownership changes.
-  Recitals recitals = 4;
-  // The executed contract.
-  Contract contract = 5;
-  // The contract signatures
-  SignatureSet signatures = 6;
-  // The address of the notary (ie the broadcaster of this message).
-  bytes invoker = 7;
-}
-
-// P8eMsgMemorializeP8eContractResponse is the out for a MemorializeP8eContract call
-message P8eMsgMemorializeP8eContractResponse {
-  option deprecated = true;
-}
-
 // The rest of the messages defined in here were copied from the old provenance repo.
 // Changes:
 // * locations of types were updated, e.g. types.PartyType became just PartyType
