+++ proto/provenance/metadata/v1/tx.proto
@@ -5,6 +5,7 @@ import "gogoproto/gogo.proto";
 import "provenance/metadata/v1/scope.proto";
 import "provenance/metadata/v1/contract.proto";
 import "provenance/metadata/v1/specification.proto";
+import "provenance/metadata/v1/p8e/p8e.proto";
 
 import "cosmos/tx/signing/v1beta1/signing.proto";
 
@@ -47,6 +48,12 @@ service Msg {
   rpc AddRecordSpecification(MsgAddRecordSpecificationRequest) returns (MsgAddRecordSpecificationResponse);
   // DeleteRecordSpecification deletes a record specification
   rpc DeleteRecordSpecification(MsgDeleteRecordSpecificationRequest) returns (MsgDeleteRecordSpecificationResponse);
+
+  // P8EMemorializeContract records the results of a P8e contract execution as a session and set of records in a scope
+  // It only exists to help facilitate the transition. The MemorializeContract is the one to use now.
+  rpc P8eMemorializeContract(MsgP8eMemorializeContractRequest) returns (MsgP8eMemorializeContractResponse) {
+    option deprecated = true;
+  };
 }
 
 // MsgMemorializeContractRequest is a request from an execution environment to record results of a contract
@@ -283,3 +290,32 @@ message MsgDeleteRecordSpecificationRequest {
 
 // MsgDeleteRecordSpecificationResponse from a delete record specification request
 message MsgDeleteRecordSpecificationResponse {}
+
+// MsgP8eMemorializeContractRequest is the input for a P8eMemorializeContract call
+message MsgP8eMemorializeContractRequest {
+  option deprecated                   = true;
+  option (gogoproto.equal)            = false;
+  option (gogoproto.goproto_stringer) = false;
+  option (gogoproto.stringer)         = false;
+  option (gogoproto.goproto_getters)  = false;
+
+  // The scope id of the object being add or modified on blockchain.
+  string scope_id = 1;
+  // The uuid of the contract execution.
+  string group_id = 2;
+  // Unique identifier for determining contract/session execution instance
+  string execution_id = 3;
+  // The new recitals for the scope.  Used in leu of Contract for direct ownership changes.
+  provenance.metadata.v1.p8e.Recitals recitals = 4;
+  // The executed contract.
+  provenance.metadata.v1.p8e.Contract contract = 5;
+  // The contract signatures
+  provenance.metadata.v1.p8e.SignatureSet signatures = 6;
+  // The address of the notary (ie the broadcaster of this message).
+  bytes invoker = 7;
+}
+
+// MsgP8eMemorializeContractResponse is the output for a P8eMemorializeContract call
+message MsgP8eMemorializeContractResponse {
+  option deprecated = true;
+}
