+++ x/metadata/types/scope.go
@@ -33,6 +33,17 @@ func NewScope(
 	}
 }
 
+// EmptyScope creates a new empty Scope.
+func EmptyScope() *Scope {
+	return &Scope{
+		ScopeId:           MetadataAddress{},
+		SpecificationId:   MetadataAddress{},
+		Owners:            []Party{},
+		DataAccess:        []string{},
+		ValueOwnerAddress: "",
+	}
+}
+
 // ValidateBasic performs basic format checking of data within a scope
 func (s *Scope) ValidateBasic() error {
 	prefix, err := VerifyMetadataAddressFormat(s.ScopeId)
@@ -109,6 +120,17 @@ func NewSession(name string, sessionID, contractSpecification MetadataAddress, p
 	}
 }
 
+// EmptySession creates a new empty Session.
+func EmptySession() *Session {
+	return &Session{
+		SessionId:       MetadataAddress{},
+		SpecificationId: MetadataAddress{},
+		Parties:         []Party{},
+		Name:            "",
+		Audit:           nil,
+	}
+}
+
 // ValidateBasic performs basic format checking of data within a scope
 func (s *Session) ValidateBasic() error {
 	prefix, err := VerifyMetadataAddressFormat(s.SessionId)
@@ -157,6 +179,17 @@ func NewRecord(name string, sessionID MetadataAddress, process Process, inputs [
 		Process:   process,
 		Inputs:    inputs,
 		Outputs:   outputs,
+	}
+}
+
+// EmptyRecord creates a new empty Record.
+func EmptyRecord() *Record {
+	return &Record {
+		Name:      "",
+		SessionId: MetadataAddress{},
+		Process:   *EmptyProcess(),
+		Inputs:    []RecordInput{},
+		Outputs:   []RecordOutput{},
 	}
 }
 
@@ -291,6 +324,15 @@ func NewProcess(name string, processID isProcess_ProcessId, method string) *Proc
 		Name:      name,
 		ProcessId: processID,
 		Method:    method,
+	}
+}
+
+// EmptyProcess creates a new empty Process.
+func EmptyProcess() *Process {
+	return &Process{
+		ProcessId: nil,
+		Name:      "",
+		Method:    "",
 	}
 }
